CE-2785 Move rendering subtitle entry point to View class
